+++ test/main_test.rb
@@ -2,10 +2,11 @@
 require 'minitest/autorun'
 require 'open3'
 require 'rbnacl'
-require '../src/Main.rb'
+require_relative '../src/Asd_illust_req.rb'
 
 class MainTest < Minitest::Test
     def test_sample
+        asd = Asd_illust_req.new
         $bot.send_message('aaa', '')
         assert_equal message.respond , a
     end
